move to dto package
