--- src/main/java/com/threeNerds/basketballDiary/mvc/dto/TeamDTO.java
@@ -1,20 +1,0 @@
-package com.threeNerds.basketballDiary.mvc.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class TeamDTO {
-
-    private String leaderId;
-
-    private String teamName;
-
-    private String hometown;
-
-    private String introduction;
-
-    private String sidoCode;
-
-    private String sigunguCode;
-}
